requirement.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.8
Django==2.2.9
django-auth-ldap==2.1.0
django-cors-headers==3.2.1
django-rest-swagger==2.1.2
djangorestframework==3.9.2
idna==2.8
importlib-metadata==1.3.0
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
openapi-codec==1.3.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.5
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.7.0
python-ldap==3.2.0
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.18
wcwidth==0.1.7
zipp==0.6.0
